pyproject.toml
[tool.poetry]
name = "hazel-db"
version = "0.1.2"
description = "A library which eases integrating SQLAlchemy into a project"
authors = ["Abdul-Hakeem Shaibu <hkmshb@gmail.com>"]
keywords = ["hazel", "sqlalchemy", "database", "db"]
license = "BSD-2-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
sqlalchemy = "^1.2"
"zope.sqlalchemy" = "^1.1"

[tool.poetry.dev-dependencies]
black = "^19.3b0"
pytest = "^3.10"
pytest-cov = "^2.6"
tox = "^3.7"

[tool.black]
line-length = 79
skip-string-normalization = true
py36 = false
exclude = '''
/(
    \.git
  | \.tox
  | \.venv
  | dist
  | build
  | docs
)/
'''

# This next section only exists for people that have their editors
# automatically call isort, black already sorts entries on its own when run.
[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
line_length = 79

[tool.tox]
legacy_tox_ini = """
[tox]
skipsdist = true
envlist = lint, py37

[testenv]
passenv=HOME USER
whitelist_externals =
    bash
    poetry
skip_install = true
commands =
    poetry install -v
    poetry run py.test --cov-report term-missing --cov=hazel_db tests

[testenv:lint]
whitelist_externals =
    bash
    poetry
skip_install = true
deps =
    pylint
commands =
    poetry install -v
    pylint --rcfile={toxinidir}/.pylintrc {toxinidir}/src/hazel_db
"""